Refactor(shared): S3 유틸리티 추출 및 Spot/Checkin prefix 분리
- DateTimeUtils, S3FileUtils로 S3 키 생성 및 시간 파싱 공통 로직 추출
- ObjectStorageService, S3UrlGenerator를 shared.infrastructure.objectstorage로 이동
- SpotService, ImageService에서 S3FileUtils.buildKey(...) 및 유틸리티 기반 삭제 로직 사용
- dev/local 환경에 spot-base-path, checkin-base-path 설정 추가
- AuthService에서 공통 S3 인프라 패키지를 참조하도록 정리
