+++ src/main/java/com/gemmap/gemmap/shared/infrastructure/objectstorage/ObjectStorageService.java
@@ -1,4 +1,4 @@
-package com.gemmap.gemmap.image.infrastructure.objectstorage;
+package com.gemmap.gemmap.shared.infrastructure.objectstorage;
 
 import com.gemmap.gemmap.shared.exception.ErrorCode;
 import com.gemmap.gemmap.shared.exception.CommonException;
