+++ src/test/java/com/gemmap/gemmap/image/application/service/ImageServiceTest.java
@@ -4,8 +4,8 @@ import com.gemmap.gemmap.image.domain.entity.Image;
 import com.gemmap.gemmap.shared.exception.CommonException;
 import com.gemmap.gemmap.image.domain.repository.ImageRepository;
 import com.gemmap.gemmap.shared.config.s3.S3Properties;
-import com.gemmap.gemmap.image.infrastructure.objectstorage.ObjectStorageService;
-import com.gemmap.gemmap.image.infrastructure.objectstorage.S3UrlGenerator;
+import com.gemmap.gemmap.shared.infrastructure.objectstorage.ObjectStorageService;
+import com.gemmap.gemmap.shared.infrastructure.objectstorage.S3UrlGenerator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
